requirement.txt
asgiref==3.4.1
attrs==24.2.0
bcrypt==3.2.0
blinker==1.4
Brotli==1.0.9
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.0.4
colorama==0.4.6
crashtest==0.4.1
cryptography==3.4.8
distlib==0.3.9
Django==3.2.25
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
Flask==2.0.3
frida==16.5.7
frida-tools==13.6.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
IP2Location==8.10.4
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jsonschema==4.17.3
kaitaistruct==0.9
keyring==24.3.1
ldap3==2.9.1
lief==0.15.1
MarkupSafe==3.0.2
mitmproxy==7.0.4
mitmproxy-windows==0.4.1
mitmproxy_rs==0.4.1
more-itertools==10.5.0
MouseInfo==0.1.3
msgpack==1.0.8
packaging==24.1
passlib==1.7.4
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.2
platformdirs==3.11.0
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.6.0
prompt_toolkit==3.0.48
protobuf==3.18.3
psutil==6.1.0
ptyprocess==0.7.0
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.21.0
pydivert==2.1.0
pyelftools==0.31
PyGetWindow==0.0.9
Pygments==2.18.0
pylsqpack==0.3.18
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyproject-api==1.8.0
pyproject_hooks==1.2.0
PyRect==0.2.0
pyrsistent==0.20.0
PyScreeze==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32-ctypes==0.2.3
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
tomlkit==0.13.2
tornado==6.4.1
tox==4.23.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
urwid==2.1.2
urwid-mitmproxy==2.1.2.1
virtualenv==20.27.1
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.1.2
wsproto==1.0.0
zipp==3.20.2
zstandard==0.15.2
